refactor(customerio-plugin): refactor customerio to match new plugin setup
