--- packages/analytics-plugin-customerio/lib/utils/inBrowser.js
@@ -1,2 +1,0 @@
-const inBrowser = typeof window !== 'undefined'
-export default inBrowser
